// src/types/CarInfo.ts
export interface CarInfo {
  registrationNumber: string;
  taxStatus: string;
  taxDueDate: string;
  motStatus: string;
  make: string;
  yearOfManufacture: number;
  engineCapacity: number;
  co2Emissions: number;
  fuelType: string;
  markedForExport: boolean;
  colour: string;
  typeApproval: string;
  revenueWeight: number;
  euroStatus: string;
  dateOfLastV5CIssued: string;
  motExpiryDate: string;
  wheelplan: string;
  monthOfFirstRegistration: string;
}